pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "murmur"
version = "0.1.0"
description = "Embed images into audio spectrograms, steganography for the frequency domain"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
requires-python = ">=3.10"
authors = [
    {name = "Zaki Pedio aka R4tz"}
]
keywords = ["steganography", "audio", "spectrogram", "cli"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Topic :: Multimedia :: Sound/Audio",
]
dependencies = [
    "numpy>=1.24",
    "Pillow>=10.0",
    "qrcode[pil]>=7.4",
]

[project.optional-dependencies]
svg = ["cairosvg>=2.7"]
decode = ["pyzbar>=0.1.9"]
dev = ["pytest>=8.0", "pytest-cov>=5.0"]

[project.scripts]
murmur = "murmur.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["murmur*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
